% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cjbColor}
\alias{cjbColor}
\alias{darken}
\alias{lighten}
\alias{add.alpha}
\alias{assignCHMcolor}
\title{Color related functions and variables}
\usage{
darken(color, factor = 1.4)

lighten(color, factor = 1.4)

add.alpha(color, alpha = 0.7)

assignCHMcolor(df, colist = comp_hm_colist_full, columns = NULL)
}
\arguments{
\item{color}{a character or vector of color}

\item{factor}{a numeric. To what extend you want to change the color}

\item{alpha}{a numeric. The alpha level.}

\item{df}{dataframe to be plotted}

\item{colist}{list of named color vector. By default is comp_hm_colist_full}

\item{columns}{if only selected columns to be used for generating the color, supply a vector of column names here.}
}
\value{
\itemize{
 \item darken,lighten: a character or vector of the changed color
 \item assignCHMcolor: a list of vectors (discrete colors) or functions (for continuous colors)
 }
}
\description{
A group of color related functions and variables.\cr
\itemize{
  \item darken,lighten: darken or lighten colors. Credit to \url{https://gist.github.com/Jfortin1}
  \item add.alpha: add alpha to color vector. Credit to \url{https://magesblog.com/post/2013-04-30-how-to-change-alpha-value-of-colours-in/}
  \item assignCHMcolor: produce a color list for complexheatmap from predefined base colors.
  }
}
\section{Predefined Colors}{

These variables will be loaded to the global environment and might overwrite existing colors.
\itemize{
  \item colpal: vector of 6 for groups
  \item colpal2: c("deepskyblue3","ivory","coral") #for heatmap
  \item comp_hm_colist_full: for complex heatmap related color
  \item colonco: for mutations on oncoprint plot
}
}

