% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexHeatmapPlots.R, R/genomicRanges.R
\name{compHeatmap}
\alias{compHeatmap}
\alias{subtypeChaHM_col}
\alias{compHeatmap}
\alias{olpByPerc}
\title{Complex heatmap plots}
\usage{
subtypeChaHM_col(df, subtype, subtype.col = NULL, cha, topbar,
  order_by = topbar, colist = comp_hm_colist_full, pic, outp, w, h)

olpByPerc(gr1, gr2, thres)
}
\arguments{
\item{df}{dataframe}

\item{subtype}{character of the name of column in the df, containing subtype info}

\item{subtype.col}{named color for the subtypes. if NULL, assign automatically.}

\item{cha}{vector of characters to be plotted}

\item{topbar}{character of the name of a numeric column with which a bar will be plotted on the top}

\item{order_by}{character of the name of column to order the heatmap}

\item{colist}{the comprehensive color list.\cr By default it is comp_hm_colist_full loaded in this package}

\item{pic}{function of plotting device, eg png, pdf, etc.}

\item{outp}{character of output folder and file name}

\item{w, h}{width and height of the output plot}

\item{df}{dataframe}

\item{subtype}{character of the name of column in the df, containing subtype info}

\item{cha}{vector of characters to be plotted}

\item{topbar}{character of the name of a numeric column with which a bar will be plotted on the top}

\item{order_by}{character of the name of column to order the heatmap}

\item{pic}{function of plotting device, eg png, pdf, etc.}

\item{outp}{character of output folder and file name}

\item{w, h}{width and height of the output plot}

\item{colist}{the comprehensive color list.\cr By default it is comp_hm_colist_full loaded in this package}
}
\value{
a plot saved in desinated path

a plot saved in desinated path
}
\description{
A group of functions for complexheatmap plots using simplied single functions.
\itemize{
  \item subtypeChaHM_col: Plot complexheatmap of certain characteristics (Age,Stage,etc), sorting by subtype, using specified color.
}

A group of functions for genomic ranges
\itemize{
  \item subtypeChaHM_col: Plot complexheatmap of certain characteristics (Age,Stage,etc), sorting by subtype, using specified color.
}
}
\seealso{
\code{\link[ComplexHeatmap]{Heatmap}}

\code{\link[ComplexHeatmap]{Heatmap}}
}
