% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_Pathway.R
\name{get_sam4grp}
\alias{get_sam4grp}
\alias{get_compdf}
\alias{fgseaWrap}
\alias{get_deg_EdgeR}
\alias{get_deg_Deseq2}
\alias{DEG_fgsea_Wrap}
\title{A group of functions for Differential gene expression and pathway analysis after that.}
\usage{
get_sam4grp(metadf, feat.coln, group1, group2, sample.coln, sep.grp = "+")

get_compdf(metadf, feat.coln, sample.coln, comp = "1vA", sep.grp = "+")

fgseaWrap(
  fn.gmt,
  ID.gene,
  stat.gene,
  nperm = 1000,
  minSize = 20,
  maxSize = 500,
  n.top = 10,
  fn.outp = NULL,
  dir.outp = NULL,
  outp.return = F
)

get_deg_EdgeR(
  countMat,
  name.feature,
  name.group1,
  name.group2,
  samID.grp1,
  samID.grp2,
  f.design = NULL,
  metadf = NULL,
  samID.coln = NULL,
  deseq.coln.trans = T,
  df.gene.anno = NULL,
  dir.outp = NULL,
  fn.outp = NULL
)

get_deg_Deseq2(
  countMat,
  name.feature,
  name.group1,
  name.group2,
  samID.grp1,
  samID.grp2,
  f.design = NULL,
  metadf = NULL,
  samID.coln = NULL,
  outpdir = NULL
)

DEG_fgsea_Wrap(
  countMat,
  name.feature,
  name.group1,
  name.group2,
  samID.grp1,
  samID.grp2,
  fn.gmt,
  f.design = NULL,
  metadf = NULL,
  samID.coln = NULL,
  outpdir = NULL
)
}
\description{
\itemize{
  \item get_sam4grp: generate tibble of library IDs according to the groups to be compared for DEG analysis
  \item get_compdf: wrapper using get_sam4grp to generate all pairwise or one-vs-all comparisons
  \item get_deg_EdgeR: generate DEG output with EdgeR
  \item get_deg_Deseq2: generate DEG output with DESeq2
  \item fgseaWrap: wrapper to run fgsea
  \item DEG_fgsea_Wrap: wrapper to run DESeq2 and fgsea
 }
}
