% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizes.R
\name{discretizes}
\alias{discretizes}
\alias{discretize}
\alias{trisect}
\title{Discretize a continuous vector}
\usage{
discretize(x, cutoff, label = NULL, include.lowest = TRUE, summary = FALSE)

trisect(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{cutoff}{numeric value (single or vector) as cutting point(s) (Don't include min or maximum),
or a function generating one cutting points (median, mean, etc) or multiple cutting points which include (quantile).}

\item{label}{change the default output region labels to customized labels}

\item{include.lowest}{see \code{\link[base]{cut}} parameter \emph{include.lowest}}

\item{summary}{when TRUE, message a summary of the count of each discrete items.}
}
\description{
A group of functions to divide continous numeric vectors into discret parts.
Similar to \code{\link[stats]{quantile}} then \code{\link[base]{cut}} but easier to use.
\itemize{
  \item discretize: cut a vector by a function or by certain cutoff value using cut.\cr
    Return a factor by replacing the original ones with the discrete value regions (<=xxx, xxx-xxx, >xxx).\cr
    \emph{NOTE: NAs are automatically removed}
  \item trisect: quantile a vector into three equal parts. To be used with discretize: \code{discretize(x,trisect)}
}
}
\examples{
discretize(rnorm(20),trisect,summary=TRUE)
discretize(rnorm(20),mean,label=c("low","high"),summary=TRUE)
}
