% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneIDs.R
\name{geneIDs}
\alias{geneIDs}
\alias{EN2other}
\title{Annotate gene symbol, geneID or EnsemblID etc}
\usage{
EN2other(x)
}
\arguments{
\item{df, x, y}{dataframe}

\item{tum}{TRUE for tumor and FALSE for normal}
}
\value{
dataframe
}
\description{
FIXME
}
\details{
A group of functions to annotate geneIDs
\itemize{
  \item del_dupcol: remove duplicated columns by name.
  \item del_alldupcol: remove columns that are not unique by name
  \item subcol_tcgaTnN: Subset tumor/normal from tcgacount
  \item subset_tcgasam: Subset one dataframe according to another, by colnames of patient name.
}
}
