% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneLengths.R
\name{geneLengths}
\alias{geneLengths}
\alias{hg19ExonLengths}
\alias{hg19cdsLengths}
\title{Retrieve gene lengths}
\usage{
hg19ExonLengths(symbols)

hg19cdsLengths(symbols)
}
\arguments{
\item{symbols}{vector of gene symbols. Will be converted into character internally by the function.}
}
\value{
a vector of lengths
}
\description{
A group of functions to gain length of genes, exons and cds.
\itemize{
  \item hg19ExonLengths: Gain total exon lengths (bp)
  \item hg19cdsLengths: Gain total cds lengths (bp)
}
}
\seealso{
Adapted from \url{https://www.biostars.org/p/62583/}
}
