% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{getPvalue}
\alias{getPvalue}
\alias{p_fish.chi.t}
\alias{p_aov.t}
\alias{p_krus}
\alias{p_lm}
\title{Returning only the p values}
\usage{
p_fish.chi.t(df, v1, v2, alt = "two.sided", ws = 2e+06)

p_aov.t(df, v1, v2)

p_krus(df, v1, v2)

p_lm(df, v1, v2)
}
\arguments{
\item{df}{numeric dataframe}

\item{v1, v2}{character of the column name in the df}

\item{alt}{passing to alternative in \code{\link[stats]{fisher.test}}, should be one of "two.sided", "greater" or "less". You can specify just the initial letter. Only used in the 2 by 2 case.}

\item{ws}{passing to workspace size in \code{\link[stats]{fisher.test}},  integer specifying the size of the workspace used in the network algorithm. In units of 4 bytes.}
}
\description{
A group of functions to return only the p values from various of tests.
\itemize{
  \item p_fish.chi.t: Try first use \code{\link[cjbmisc]{fish.t}} to get p value,\cr
  if failed (usually due to too many catergories), then use \code{\link[cjbmisc]{chisq.t}} with popup warning.
  \item p_aov.t: p-value from \code{\link[stats]{aov}}
  \item p_krus: p-value from \code{\link[stats]{kruskal.test}}
  \item p_lm: p-value of the coefficient from the univariate \code{\link[stats]{lm}}
}
}
\seealso{
\code{\link{chisq.t}}, \code{\link{fish.t}}, \code{\link[stats]{fisher.test}},\code{\link[stats]{chisq.test}},
\code{\link[stats]{aov}},\code{\link[stats]{kruskal.test}},\code{\link[stats]{lm}}
}
