% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairscjb.R
\name{ggpairs_custom}
\alias{ggpairs_custom}
\title{Plot pairwise correlations of multiple features}
\usage{
ggpairs_custom(ggdf, feat.plot = NULL, ...)
}
\arguments{
\item{ggdf}{dataframe with rows of samples and columns of features}

\item{feat.plot}{character vector of the features to be plotted. If NULL then all columns will be plotted}

\item{...}{pass to \code{\link[GGally]{ggpairs}}}
}
\value{
a "gg" "ggmatrix" plotable and ggsavable object
}
\description{
Only a beta version
}
