% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurvpWrap.R
\name{ggsurvpWrap}
\alias{ggsurvpWrap}
\title{Survival plot with ggsurvplot
A warpper to plot survival plots
to further customize the plot,}
\usage{
ggsurvpWrap(survdf, fac, title = NULL, col = comp_hm_colist_full$disc,
  risk.table = F, discretize.method = median, xlim = c(0, 120),
  break.time.by = 60, textsize = 10, pval.method = T, ...)
}
\arguments{
\item{survdf}{dataframe with "OS_Month" (character or numeric) and "OS_Status" (numeric of 0,1; or factor with Alive as base level)}

\item{fac}{character of the column name in the survdf to stratify the patients}

\item{col}{vector of colors}

\item{discretize.method}{if \emph{fac} is numeric, how to discretizeit. Can be funcitons (Default median) or cutoff points (n-1 points)}

\item{...}{arguments to ggsurvplot}
}
\value{
A object of class ggsurvplot
}
\description{
Survival plot with ggsurvplot
A warpper to plot survival plots
to further customize the plot,
}
\seealso{
\code{\link[survminer]{ggsurvplot}}
}
