% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{htmap}
\alias{htmap}
\title{Heatmaps
A group of functions for heatmaps
FIXME}
\usage{
htmap(countdata, coldata, title, breaks = seq(-1.8, 1.8, length.out = 201)^3 +
  0.5)
}
\description{
Heatmaps
A group of functions for heatmaps
FIXME
}
