% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscPlots.R
\name{miscPlots}
\alias{miscPlots}
\alias{perc_barplot}
\alias{violin}
\alias{boxjitter}
\alias{plot_lm}
\title{Miscellaneous plotting functions}
\usage{
perc_barplot(
  ggdf,
  x,
  y,
  title = NULL,
  col = comp_hm_colist_full$disc,
  xlab = x,
  ylab = y,
  width = 0.9,
  aspect.ratio = 1.4,
  p.label = "p.format",
  p.symnum.args = list(),
  show.count = TRUE,
  anno.textsize = 4,
  anno.vjust = 0.04,
  border = TRUE,
  plot.it = FALSE,
  highlight.signif = TRUE,
  highlight.signif.col = "orchid1",
  signif.cutoff = 0.05,
  ...
)

violin(
  ggdf,
  x,
  y,
  test = "wilcox.test",
  title = NULL,
  col = NULL,
  xlab = x,
  ylab = y,
  width = 0.9,
  aspect.ratio = 1.4,
  show.mean = TRUE,
  show.count = TRUE,
  p.label = NULL,
  p.symnum.args = list(),
  anno.vjust = 0.1,
  anno.textsize = 4,
  border = TRUE,
  plot.it = FALSE,
  highlight.signif = TRUE,
  highlight.signif.col = "orchid1",
  signif.cutoff = 0.05,
  plot.type = "violin",
  box.notch = FALSE,
  ...
)

boxjitter(...)

plot_lm(
  ggdf,
  x,
  y,
  title = NULL,
  col = "#00DD006F",
  show.anno = TRUE,
  cor.method = "spearman",
  plot.it = TRUE,
  highlight.signif = TRUE,
  highlight.signif.col = "orchid1",
  signif.cutoff = 0.05,
  ...
)
}
\arguments{
\item{ggdf, df}{dataframe with rows of samples and columns of features}

\item{x, y}{character of the name of columns to be plotted}

\item{title}{plot title}

\item{col}{For perc_barplot & violin: named vector of color (as long as the values are contained in the names), or unnamed vector, then color would be assigned based on the order of the levels.\cr For plot_lm: one color with alpha.}

\item{xlab, ylab}{axis label}

\item{width}{perc_barplot & violin: with of the bar or violin}

\item{aspect.ratio}{perc_barplot & violin:aspect ratio of the final plot}

\item{p.label, p.symnum.args}{change the label style ("p.signif" or "p.format") of the significance. See label and symnum.args from \code{\link[ggpubr]{stat_compare_means}}}

\item{show.count}{perc_barplot & violin:whether to show the total count of each catergory}

\item{border}{perc_barplot,violin: whether add a solid border to the plot}

\item{plot.it}{if FALSE, no plot printed, only return the plot object}

\item{highlight.signif, highlight.signif.col, signif.cutoff}{Whether to highlight (highlight.signif) a significant plot with a color border (highlight.signif.col) according to the p-value cutoff (signif.cutoff).\cr
pvalue calculationg method: perc_barplot-\code{\link{p_fish.chi.t}}; violin-\code{\link{p_krus}}; plot_lm-from the univariate \code{\link[stats]{lm}}}

\item{...}{in perc_barplot, violin, plot_lm: pass to \code{\link[ggplot2]{theme}}}

\item{show.mean}{violin:whether to show the mean with a green dot}

\item{box.notch}{violin/boxjitter: whether show notch style box plot}

\item{show.anno}{plot_lm: whether to show the R2, slope and slope p value}

\item{cor.method}{plot_lm:the correlation method used for the annotation, pass to \code{\link[stats]{cor}}}

\item{anno.hjust, anno.textsize}{violin:change the position and text size of the significance/count values}

\item{pch}{plot_lm:dot style}
}
\value{
A ggplot object
}
\description{
A group of functions for miscellaneous plots. These are the foundation of the plot_feat_subtype and ggpairs_custom\cr
You may adjust via trial&error for the anno.hjust and anno.textsize parameter if the p-value text size and position are not good enough.
\itemize{
  \item perc_barplot: Percentage barplot for two discrete values, showing the proportion of y in each x. fish or chisq test p value shown.
  \item violin: violin plot for one discrete and one continuous values. pvalue shown.
  \item boxjitter: same as violin plot, but show the data with a box plot and jitter points.
  \item plot_lm: dot plot with lm regression line for two continuous values. Also show the correlation, R2, p value of slope
  \item violin_mul: plot multiple violin plot, but only the significant ones.
}
}
