% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprintPlot.R
\name{oncoprint}
\alias{oncoprint}
\alias{maf2oncoprintdf}
\title{Oncoprint plots}
\usage{
maf2oncoprintdf(inpmaf, gene = NULL, ordergene = FALSE, coding = TRUE,
  noSilent = FALSE, mergeVariant = TRUE, subtypeOrder = NULL)
}
\arguments{
\item{inpmaf}{dataframe of the input maf, should have the columns of Variant_Classification, Hugo_Symbol, and Tumor_Sample_Barcode}

\item{gene}{vector of characters for the symbols of genes to be plotted. Default all genes.}

\item{ordergene}{boolean for whether the genes should be orderd by the order of the gene list}

\item{coding}{boolean for whether to subset only the coding mutations}

\item{noSilent}{boolean for whether to remove the silent mutations}

\item{mergeVariant}{boolean for whether to merge the "Del and Ins" to "Indel" and various mutations to "Translation_Start_Site"}

\item{subtypeOrder}{named character of the subtype info of the patients. If provided, the oncodf would be ordered by this subtype info.}

\item{outp}{character of output folder and file name}

\item{w, h}{width and height of the output plot}

\item{colist}{the comprehensive color list.\cr By default it is comp_hm_colist_full loaded in this package}
}
\value{
a plot saved in desinated path
}
\description{
A group of functions for oncoprint plots using simplied single functions.
\itemize{
  \item maf2oncoprintdf: Input a maf, generate the dataframe of gene mutation types for each patients as the input for oncoprint plot.
}
}
\seealso{
\code{\link[ComplexHeatmap]{oncoPrint}}
}
