% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlots.R
\name{PCAPlots}
\alias{PCAPlots}
\alias{pca}
\alias{pca_multi}
\alias{pca_bg}
\title{PCA Plots}
\usage{
pca(
  df,
  A,
  B = NULL,
  title = "PCA",
  col = colpal,
  nameA = "subtypeA",
  nameB = "subtypeB",
  point.size = 4,
  shapes = c(16:18, 15),
  plot.it = TRUE,
  return.data = FALSE
)

pca_multi(
  df,
  subtype = NULL,
  n = 6,
  col = colpal,
  title = NULL,
  plot.it = TRUE
)

pca_bg(
  df,
  subtype,
  col = colpal,
  legend.name = NULL,
  point.size = 6,
  text.size = 26,
  mar = 15,
  plot.it = TRUE
)
}
\arguments{
\item{df}{numeric dataframe with rows of samples and columns of features}

\item{A, B, subtype}{character vector of the subtypes of samples}

\item{col}{color to be used for each subtype.}

\item{plot.it}{if FALSE, no plot printed, only return the plot object}

\item{n}{for pca_multi, number of PCs to be plotted}

\item{mar}{for pca_bg, margin between the points to the frame.}

\item{"+"}{use + to add ggplot layers for more customization}
}
\value{
Print the plot and return the plot
\itemize{
  \item pca: a ggplot object
  \item pca_multi: a plot object
}
}
\description{
A group of functions for PCA plots.
\itemize{
  \item pca: PCA plot with one or two group info.
  \item pca_multi: Multiple PCA plots for top PCs.
  \item pca_bg: PCA plot with color background shading for each subtype
}
}
\seealso{
\code{\link[ggplot2]{autoplot}} and \code{\link{prcomp}}
}
