% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrExplorationPlots.R
\name{plot_corr}
\alias{plot_corr}
\alias{plot_corr_categorical}
\alias{plot_corr_numeric}
\title{Plots to explore the correlations between features}
\usage{
plot_corr_categorical(
  plotdf,
  x.coln,
  y.coln = NULL,
  cat.num.test = "kruskal.test",
  cat.cat.test = "both",
  plot.it = F,
  plot.nrow = NULL,
  plot.ncol = NULL,
  signif.cutoff = 0.05,
  plot.signif.only = F,
  seed = 999,
  outpdir = NULL,
  plot.w = 6.5,
  plot.h = 7.5,
  ...
)

plot_corr_numeric(
  plotdf,
  x.coln,
  y.coln,
  cat.num.test = "kruskal.test",
  num.num.test = "spearman",
  plot.it = F,
  plot.nrow = NULL,
  plot.ncol = NULL,
  signif.cutoff = 0.05,
  plot.signif.only = F,
  seed = 999,
  outpdir = NULL,
  plot.w = 6.5,
  plot.h = 7.5,
  ...
)

plot_corr(
  plotdf,
  x.coln,
  y.coln = NULL,
  cat.num.test = "kruskal.test",
  cat.cat.test = "both",
  num.num.test = "spearman",
  plot.it = F,
  plot.nrow = NULL,
  plot.ncol = NULL,
  signif.cutoff = 0.05,
  plot.signif.only = F,
  seed = 999,
  outpdir = NULL,
  plot.w = 6.5,
  plot.h = 7.5,
  ...
)
}
\arguments{
\item{plotdf}{dataframe with rows of samples and columns of features.}

\item{x.coln}{column name of the x axis of the plot}

\item{y.coln}{character vector of the column names of features to be plotted as y axis}

\item{cat.num.test}{the significance test to be used for categorical vs numerical variables. Use the name of the r basic tests (Default "kruskal.test").}

\item{cat.cat.test}{the significance test to be used for categorical vs categorical variables. Should be "fisher","chi" or "both"(Default)}

\item{plot.it}{Whether to plot it out (T/F)}

\item{plot.nrow, plot.ncol}{The number of rows and columns in the combined plot}

\item{plot.signif.only}{whether to plot only the significant items}

\item{outpdir}{If not NULL, save all plots and pvalues (as table) to the outpdir}

\item{plot.w, plot.h}{Width and height of each individual plot}

\item{...}{pass to \code{\link[ggstatsplot]{ggbarstats}} , \code{\link[ggstatsplot]{ggbetweenstats}} ,\code{\link[ggstatsplot]{ggscatterstats}}}

\item{num.num.test}{the significance test to be used for numerical vs numerical variables. Should be "spearman"(Default), "pearson", "kendall", or "lm"(using the pvalue of the independent variable in lm).}
}
\value{
For plot_corr, List of returns from plot_corr_numeric or plot_corr_categorical. For plot_corr_categorical and plot_corr_numeric: List of two: "plot" of ggarrange object which arrange all plot into one, and "pvalues" of a named vector.
}
\description{
Plot the correlations focusing on a variable x vs all the rest
of the variables.
}
