% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceNames.R
\name{replaceNames}
\alias{replaceNames}
\title{Replace the names}
\usage{
replaceNames(data, c, n)
}
\arguments{
\item{data}{a vector or a dataframe}

\item{c}{names before change (vector names or column names)}

\item{n}{names after change}
}
\value{
the dataframe or vector which using new names (not affect the data)
}
\description{
Replace the names from vector or dataframe
}
\examples{
v <- setNames(c("a","b","c"),c("a","b","c"))
replaceNames(v,c("b","c"),c("d","f"))
}
