% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq.R
\name{RNAseq}
\alias{RNAseq}
\alias{normSizeF}
\alias{nmfPre}
\alias{nmfCluster}
\alias{write_gct}
\alias{write_cls}
\title{A group of functions for RNAseq related analysis.}
\usage{
normSizeF(countdata)

nmfPre(data, rank = 3:6, r = 50)

nmfCluster(data, rank = 3, r = 200)

write_gct(exprdf, gctfn, genelist = NULL)

write_cls(subtype, clsfn)
}
\description{
\itemize{
  \item normSizeF: perform DESeq2 normalization
  \item nmfPre: pre-clustering for NMF \code{\link[NMF]{nmf}} to determine the optimal rank
  \item nmfCluster: NMF \code{\link[NMF]{nmf}} clustering output the NMF object and the assigned subtype
  \item write_gct: transform a expression matrix (gene in rows with rowname) into gct format and save file with name "gctfn".
  \item write_cls: generate a cls format from a vector of subtype and save file with name "clsfn".
}
}
