% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTests.R
\name{stattests}
\alias{stattests}
\alias{fish.t}
\alias{chisq.t}
\alias{one.v.all.t}
\title{Simplified stat tests}
\usage{
fish.t(df, v1, v2, alt = "two.sided", ws = 2e+07)

chisq.t(df, v1, v2)

one.v.all.t(df, test = fisher.test, ...)
}
\arguments{
\item{df}{numeric dataframe}

\item{v1, v2}{character of the column name in the df}

\item{alt}{passing to alternative in \code{\link[stats]{fisher.test}}, should be one of "two.sided", "greater" or "less". You can specify just the initial letter. Only used in the 2 by 2 case.}

\item{ws}{passing to workspace size in \code{\link[stats]{fisher.test}},  integer specifying the size of the workspace used in the network algorithm. In units of 4 bytes.}

\item{test}{the test to be used (fisher.test, etc). the ... is passed to the test function}
}
\description{
A group of functions to quickly run certain statistical tests from a dataframe.
\itemize{
  \item fish.t: Fisher's exact test with columns in dataframe.
  \item chisq.t Chi Square Test with columns in dataframe
  \item one.v.all.t: One versus all test of contingency table, return a named p value vector
}
}
\seealso{
\code{\link[stats]{fisher.test}},\code{\link[stats]{chisq.test}},
}
