% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCols.R
\name{subsetCols}
\alias{subsetCols}
\alias{del_dupcol}
\alias{del_alldupcol}
\alias{subset_colname}
\alias{subcol_tcgaTnN}
\alias{subset_tcgasam}
\title{Subset columns from a dataframe by various criteria}
\usage{
del_dupcol(df)

del_alldupcol(df)

subset_colname(x, y, order = T)

subcol_tcgaTnN(df, tum = T)

subset_tcgasam(x, y)
}
\arguments{
\item{df, x, y}{dataframe}

\item{tum}{TRUE for tumor and FALSE for normal}
}
\value{
dataframe
}
\description{
A group of functions to subset columns from a dataframe 
\itemize{
  \item del_dupcol: remove duplicated columns by name.
  \item del_alldupcol: remove columns that are not unique by name
  \item subset_colname: Subset one dataframe according to another, by colnames. Return df has same order as the reference df
  \item subcol_tcgaTnN: Subset tumor/normal from tcgacount
  \item subset_tcgasam: Subset one dataframe according to another, by colnames of patient name. Will truncate long names to the patient barcode for you.
}
}
