% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cjb.misc}
\alias{cjb.misc}
\alias{dist_cos}
\alias{dist_euc}
\alias{get_mode}
\alias{cv}
\alias{perm}
\alias{consistSymbol}
\title{Miscellaneous functions}
\usage{
dist_cos(a, b)

dist_euc(a, b)

get_mode(v, na.rm = TRUE, multi = FALSE)

cv(x)

perm(n)

consistSymbol(a, r)
}
\arguments{
\item{a, b, x, r}{numeric vector. r for reference}

\item{v}{any vector}

\item{multi}{when TRUE, return a vector when multiple outcome are possible; when FALSE, only output the first one of them.}

\item{n}{integer}
}
\description{
A group of functions for miscellaneous usage.
\itemize{
  \item dist_cos, dist_euc: cosine and Euclidean distance of 2 vectors.\cr
  return a single numeric value
  \item get_mode, cv: get the mode and coefficient of variations of a vector.\cr
  return a single numeric value or multiple values when \emph{multi==TRUE}
  \item perm: permutation of 1:n sequence
  \item consistSymbol: Force a integer vector to use numbers match best with another vector
}
}
