% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairscjb.R
\name{ggpairs_custom}
\alias{ggpairs_custom}
\title{Plot pairwise correlations of multiple features}
\usage{
ggpairs_custom(ggdf, feat.plot = NULL, colist = comp_hm_colist_full,
  sig.col = c("white", "thistle1", "orchid1"), plot.it = FALSE,
  signif.cutoff = 0.05, return.pvalue = FALSE, ...)
}
\arguments{
\item{ggdf}{dataframe with rows of samples and columns of features}

\item{feat.plot}{character vector of the features to be plotted. If NULL then all columns will be plotted}

\item{colist}{a named list of named vector of colors to be used for each feature. Missing ones are automatically handled.}

\item{sig.col}{vector of 3 colors for nonsignif, lower and higher bound of the significance.}

\item{plot.it}{if FALSE, no plot printed, only return the object. Use ggsave(plot=p$plot) to plot.}

\item{signif.cutoff}{cutoff point for significant p-value. A colored (sig.col[3]) frame will be shown to significant plots.}

\item{return.pvalue}{whether to return a pvalue matrix. If yes, it has to plot it to generate the values (to a new device)}

\item{...}{pass to \code{\link[GGally]{ggpairs}}}
}
\value{
If return.pvalue=FALSE:"gg" "ggmatrix" plotable and ggsavable object; if TRUE: a list of two: $plot is the plot as when return.pvalue=FALSE; $p.value is the table of correlation pvalues
}
\description{
Plot a big complex correlation map of every features in a dataframe or specific features (feat.plot).
Usage: ggpairs_custom(df,plot.it=TRUE) (all features). Or p<-ggpairs_custom(df,c("Ploidy","Purity","EGFR","Smoker"))\cr
then ggsave("plot.pdf",p)
}
