% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurvpWrap.R
\name{ggsurvpWrap}
\alias{ggsurvpWrap}
\title{Survival plot with ggsurvplot
A warpper to plot survival plots
to further customize the plot,}
\usage{
ggsurvpWrap(survdf, fac, title = NULL, col = NULL, risk.table = F,
  xlim = c(0, 120), break.time.by = 60, textsize = 10, pval.method = T,
  ...)
}
\arguments{
\item{survdf}{dataframe with "OS_Month" (character or numeric) and "OS_Status" (numeric of 0,1; or factor with Alive as base level)}

\item{fac}{character of the column name in the survdf, which is a factor to stratify the patients}

\item{col}{vector of colors}

\item{...}{arguments to ggsurvplot}
}
\value{
A object of class ggsurvplot
}
\description{
Survival plot with ggsurvplot
A warpper to plot survival plots
to further customize the plot,
}
\seealso{
\code{\link[survminer]{ggsurvplot}}
}
