% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurvpWrap.R
\name{ggsurvpWrap}
\alias{ggsurvpWrap}
\title{Survival plot with ggsurvplot
A warpper to plot survival plots
to further customize the plot,}
\usage{
ggsurvpWrap(
  survdf,
  fac,
  surv.time = "OS_Month",
  surv.status = "OS_Status",
  cont.to.dis = NA,
  title = fac,
  col = comp_hm_colist_full$disc,
  risk.table = F,
  xlim = c(0, 120),
  break.time.by = 60,
  textsize = 10,
  pval.method = T,
  ...
)
}
\arguments{
\item{survdf}{dataframe with \emph{surv.time} (character or numeric) and \emph{surv.status}  (numeric of 0,1; or factor with Alive as base level)}

\item{fac}{character of the column name in the survdf to stratify the patients}

\item{surv.time, surv.status}{column name of the survival time and status in the dataframe}

\item{cont.to.dis}{if \emph{fac} is numeric, how to discretize it. Can be a funciton (Default \emph{median}) or cutoff points (n-1 points)}

\item{col}{vector of colors}

\item{xlim, break.time.by, pval.method, risk.table, ...}{arguments to \code{\link[survminer]{ggsurvplot}}}

\item{textsize}{pass to ggplot global text size}
}
\value{
A object of class ggsurvplot
}
\description{
Survival plot with ggsurvplot
A warpper to plot survival plots
to further customize the plot,
}
