% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscPlots.R
\name{miscPlots}
\alias{miscPlots}
\alias{perc_barplot}
\alias{violin}
\alias{plot_lm}
\title{Miscellaneous plotting functions}
\usage{
perc_barplot(ggdf, x, y, title = NULL, col = colpal, xlab = x, ylab = y,
  width = 0.9, aspect.ratio = 1.4, show.count = TRUE, anno.vjust = 0.04,
  ...)

violin(ggdf, x, y, test = "wilcox.test", title = NULL, col = NULL,
  xlab = x, ylab = y, width = 0.9, aspect.ratio = 1.4,
  show.mean = TRUE, show.count = TRUE, anno.vjust = 0.1,
  anno.textsize = 4, plot.it = TRUE, ...)

plot_lm(df, f, title = "", col = "#00DD006F", pch = 16, ...)
}
\arguments{
\item{ggdf, df}{dataframe with rows of samples and columns of features}

\item{x, y}{character of the name of columns to be plotted}

\item{title}{plot title}

\item{col}{named vector of color (as long as the values are contained in the names), or unnamed vector, then color would be assigned based on the order of the levels.}

\item{xlab, ylab}{axis label}

\item{width}{perc_barplot & violin: with of the bar or violin}

\item{aspect.ratio}{perc_barplot & violin:aspect ratio of the final plot}

\item{show.count}{perc_barplot & violin:whether to show the total count of each catergory}

\item{...}{in perc_barplot and violin: pass to \code{\link[ggplot2]{theme}};\cr in plot_lm: pass to \code{\link[graphics]{plot}}}

\item{show.mean}{violin:whether to show the mean with a green dot}

\item{plot.it}{if FALSE, no plot printed, only return the plot object}

\item{f}{formula, (y~x)}

\item{pch}{plot_lm:dot style}

\item{anno.hjust, anno.textsize}{violin:change the position and text size of the significance values}
}
\description{
A group of functions for miscellaneous plots.
\itemize{
  \item perc_barplot: Percentage barplot for two discrete values, showing the proportion of y in each x. fish or chisq test p value shown.
  \item violin: violin plot for one discrete and one continuous values. pvalue shown.
  \item plot_lm: dot plot with lm regression line for two continuous values. Also show R2, p value of slope
  \item violin_mul: plot multiple violin plot, but only the significant ones.
}
}
