% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypeFeaturePlot.R
\name{plot_feat_subtype}
\alias{plot_feat_subtype}
\title{Plot multiple features of subtypes}
\usage{
plot_feat_subtype(plotdf, subtype, feat.plot, cont.plot = "boxplot",
  cont.test = "wilcox.test", cont.show.mean = FALSE, show.count = TRUE,
  disc.col = comp_hm_colist_full, cont.col = comp_hm_colist_full$disc,
  disc.width = 0.8, cont.width = 0.9, cont.anno.vjust = 0.2,
  disc.anno.vjust = 0.04, aspect.ratio = 1.4)
}
\arguments{
\item{plotdf}{dataframe with rows of samples and columns of features, one column should contain the subtype info.}

\item{subtype}{column name of the subtype}

\item{feat.plot}{vector of the features to be plotted}

\item{cont.plot}{character of either "boxplot" or "violin"}

\item{cont.test}{the test to be used for the significant test for continuous variables}

\item{cont.show.mean}{boolean of whether to show the mean as a green dot for continuous variables}

\item{show.count}{whether to show the total count of each catergory}

\item{disc.col}{list of named vectors for colors of discrete variables}

\item{cont.col}{vector of colors for continuous variables}

\item{cont.anno.vjust, disc.anno.vjust}{numeric to tweak the position of the pvalue annotation on the plot}
}
\description{
Plot percentage barplot or violin/boxplot for discrete or continuous variables of each subtype.
Return a list of ggplots that can be arranged into figure panel with ggarrange.
}
